requirements.txt
streamlit==1.30.0
streamlit-authenticator==0.5.5
sqlalchemy==2.0.18
gitpython==3.1.30
pandas==2.2.3
python-dotenv==1.0.0
